+++ src/record/internal/matroska_read.cpp
@@ -744,8 +744,8 @@ k4a_result_t parse_recording_config(k4a_playback_context_t *context)
         start_offset_str >> start_offset_ns;
         if (!start_offset_str.fail())
         {
-            assert(start_offset_ns / 1000 <= UINT32_MAX);
-            context->record_config.start_timestamp_offset_usec = (uint32_t)(start_offset_ns / 1000);
+            assert(start_offset_ns / 1000 <= UINT64_MAX);
+            context->record_config.start_timestamp_offset_usec = (uint64_t)(start_offset_ns / 1000);
         }
         else
         {
@@ -1851,6 +1851,7 @@ k4a_result_t convert_block_to_image(k4a_playback_context_t *context,
 
     if (K4A_SUCCEEDED(result) && buffer != NULL)
     {
+        //TODO:NV12¸ñÊ½
         result = TRACE_CALL(k4a_image_create_from_buffer(target_format,
                                                          out_width,
                                                          out_height,
@@ -1892,6 +1893,7 @@ k4a_result_t new_capture(k4a_playback_context_t *context, block_info_t *block, k
     {
         result = TRACE_CALL(convert_block_to_image(context, block, &image_handle, context->color_format_conversion));
         k4a_capture_set_color_image(*capture_handle, image_handle);
+        
     }
     else if (block->reader == context->depth_track)
     {
